Mover as classes TripListUiState e UiState para novos arquivos e ajustar imports necessários
