+++ androidApp/src/main/java/com/douglastaquary/busaoshare/android/ui/viewModels/SearchTripViewModel.kt
@@ -3,24 +3,12 @@ package com.douglastaquary.busaoshare.android.ui.viewModels
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import co.touchlab.kermit.Logger
+import com.douglastaquary.busaoshare.android.ui.state.UiState
 import com.douglastaquary.busaoshare.model.Result
 import com.douglastaquary.busaoshare.model.Trip
 import com.douglastaquary.busaoshare.repository.SPTransAPIRepository
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.*
-
-data class TripListUiState(
-    val items: List<Trip> = emptyList(),
-    val isLoading: Boolean = false,
-    val userMessage: String? = null
-)
-
-sealed class UiState<out T : Any> {
-    object Loading : UiState<Nothing>()
-    object Empty : UiState<Nothing>()
-    data class Success<out T : Any>(val data: T) : UiState<T>()
-    data class Error(val exception: Exception) : UiState<Nothing>()
-}
 
 @ExperimentalCoroutinesApi
 class SearchTripViewModel(
